提交周计划,立即置灰

+++ app/src/main/java/et/tsingtaopad/dd/ddxt/term/select/XtTermSelectFragment.java
@@ -257,11 +257,13 @@ public class XtTermSelectFragment extends BaseFragmentSupport implements View.On
                     item.setIsSelectToCart("0");
                     imageView.setImageResource(R.drawable.icon_visit_add);
                     confirmTv.setText("确定" + "(" + selectedList.size() + ")");
+
                 } else {
                     selectedList.add(item);
                     item.setIsSelectToCart("1");
                     imageView.setImageResource(R.drawable.icon_select_minus);
                     confirmTv.setText("确定" + "(" + selectedList.size() + ")");
+
                 }
             }
         });
@@ -363,25 +365,31 @@ public class XtTermSelectFragment extends BaseFragmentSupport implements View.On
 
         // 跳转购物车Fragment
         if (TOFRAGMENT == type) {
-            // 清空购物车表数据  // 购物车表ddtype 1:协同  2:追溯
-            xtSelectService.deleteCartData("MST_TERMINALINFO_M_CART","1");
-            // 复制终端临时表
-            for (XtTermSelectMStc xtselect : selectedList) {
-                copyMstTerminalinfoMCart(xtselect);
+            if(selectedList.size()>0){
+                // 清空购物车表数据  // 购物车表ddtype 1:协同  2:追溯
+                xtSelectService.deleteCartData("MST_TERMINALINFO_M_CART","1");
+                // 复制终端临时表
+                for (XtTermSelectMStc xtselect : selectedList) {
+                    copyMstTerminalinfoMCart(xtselect);
+                }
+                // 销毁当前Fragment
+                supportFragmentManager.popBackStack();
+
+                /*Bundle bundle = new Bundle();
+                bundle.putSerializable("fromFragment", "XtTermSelectFragment");
+                XtTermCartFragment xtTermCartFragment = new XtTermCartFragment();
+                xtTermCartFragment.setArguments(bundle);
+
+                // 跳转终端购物车
+                changeHomeFragment(xtTermCartFragment, "xttermcartfragment");*/
+                changeHomeFragment(new XtTermCartFragment(), "xttermcartfragment");
+                // 购物车是否已经同步数据  false:没有  true:已同步
+                PrefUtils.putBoolean(getActivity(),GlobalValues.XT_CART_SYNC,false);// 设置需要同步
+
+                Toast.makeText(getActivity(),"终端夹添加成功",Toast.LENGTH_SHORT).show();
+            }else{
+                Toast.makeText(getActivity(),"请先添加终端",Toast.LENGTH_SHORT).show();
             }
-            // 销毁当前Fragment
-            supportFragmentManager.popBackStack();
-
-            /*Bundle bundle = new Bundle();
-            bundle.putSerializable("fromFragment", "XtTermSelectFragment");
-            XtTermCartFragment xtTermCartFragment = new XtTermCartFragment();
-            xtTermCartFragment.setArguments(bundle);
-
-            // 跳转终端购物车
-            changeHomeFragment(xtTermCartFragment, "xttermcartfragment");*/
-            changeHomeFragment(new XtTermCartFragment(), "xttermcartfragment");
-            // 购物车是否已经同步数据  false:没有  true:已同步
-            PrefUtils.putBoolean(getActivity(),GlobalValues.XT_CART_SYNC,false);// 设置需要同步
         }
     }
